Cleanup and fixes
